More endpoints done

+++ backend/endpoints/racks/controllers/deleteRack.js
@@ -1,5 +1,6 @@
 const Rack = require('../../../models/Rack'),
-   Asset = require('../../../models/Asset');
+   Asset = require('../../../models/Asset'),
+   { isValidObjectId } = require('mongoose');
 
 /**
  * @openapi
@@ -26,6 +27,8 @@ const Rack = require('../../../models/Rack'),
  *                   type: boolean
  *                 message:
  *                   type: string
+ *       '400':
+ *         description: Bad request
  *       '404':
  *         description: Not found
  *       '500':
@@ -41,20 +44,26 @@ module.exports = async (req, res) => {
    try {
       const { id } = req.params || {};
 
+      // Check id
       if (!id) {
-         return res.status(400).json({ success: false, message: 'Rack ID missing from request' });
+         return res
+            .status(400)
+            .json({ success: false, message: 'Rack ID is missing from the request' });
       }
 
-      const deletedRack = await Rack.findByIdAndDelete(id);
+      if (!isValidObjectId(id)) {
+         return res.status(400).json({ success: false, message: 'Rack ID is invalid' });
+      }
 
-      if (!deletedRack) {
+      // Find and delete the rack
+      const rack = await Rack.findByIdAndDelete(id);
+
+      if (!rack) {
          return res.status(404).json({ success: false, message: 'Rack not found' });
       }
 
-      const assets = await Asset.find({ rack: deletedRack._id });
-      const uuids = assets.map((asset) => asset.uuid);
-
-      await Asset.deleteMany({ uuid: { $in: uuids } });
+      // Delete all assets associated with the rack
+      await Asset.deleteMany({ rack: rack._id });
 
       return res.status(200).json({ success: true, message: 'Rack deleted' });
    } catch (err) {
